+++ problems/laboratory-problems/dilution_factor_diluent_numeric.py
@@ -11,10 +11,10 @@ def make_question_text(volume, df_value):
 	df_text = lab_helper_lib.format_df(lab_helper_lib.format_monospace(f"DF={df_value}"))
 	key_request = lab_helper_lib.format_key_request(f"{volume_text} with dilution factor {df_text}")
 	aliquot_text = lab_helper_lib.format_aliquot('aliquoted sample')
-	diluent_text = lab_helper_lib.format_diluent('diluent')
+	diluent_text = lab_helper_lib.format_diluent('volume of diluent')
 	question = (
 		f"<p>You are preparing {key_request}.</p>"
-		f"<p>What volume of {diluent_text} in milliliters (mL) do you add to the {aliquot_text}?</p>"
+		f"<p>What {diluent_text} in milliliters (mL) do you add to the {aliquot_text}?</p>"
 	)
 	return question