+++ problems/laboratory-problems/serial_dilution_factor_diluent_numeric.py
@@ -26,11 +26,11 @@ def question_text(volume, df1, df2):
 	volume_request = lab_helper_lib.format_key_request(volume_text)
 	aliquot_text = lab_helper_lib.format_aliquot('aliquoted sample')
 	aliquot_sample = lab_helper_lib.format_aliquot('previous diluted sample')
-	diluent_text = lab_helper_lib.format_diluent('diluent')
+	diluent_text = lab_helper_lib.format_diluent('volume of diluent')
 	question = (
 		f"<p>Using a {aliquot_sample} at {df1_request}, "
 		f"create a new dilution with a final dilution of {df2_request} and a total volume of {volume_request}.</p>"
-		f"<p>What volume of {diluent_text} in microliters (&mu;L) do you add to the {aliquot_text}?</p>"
+		f"<p>What {diluent_text} in microliters (&mu;L) do you add to the {aliquot_text}?</p>"
 	)
 	return question